pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "emailagent"
version = "1.0.0"
description = "Automated job application tracker that scans Gmail for job-related emails"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "EmailAgent Team"}
]
keywords = ["email", "gmail", "job-tracker", "automation", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Email",
    "Topic :: Office/Business",
]
dependencies = [
    "google-auth>=2.16.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.70.0",
    "openpyxl>=3.1.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "colorama>=0.4.6",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
    "flake8>=6.1.0",
]

[project.scripts]
emailagent = "cli:app"

[project.urls]
Homepage = "https://github.com/emailagent/emailagent"
Documentation = "https://github.com/emailagent/emailagent#readme"
Issues = "https://github.com/emailagent/emailagent/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "job_tracker*", "bulk_cleaner*"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.venv
    | venv
    | build
    | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=core --cov=job_tracker --cov-report=html"
